COMMERCE-1037 commerce-openapi-core-impl - Make organization of classes consistent
