+++ commerce-openapi-core-impl/src/main/java/com/liferay/commerce/openapi/core/internal/message/body/JsonMessageBodyReader.java
@@ -12,25 +12,23 @@
  * details.
  */
 
-package com.liferay.commerce.openapi.core.internal.jaxrs.json;
+package com.liferay.commerce.openapi.core.internal.message.body;
 
-import com.fasterxml.jackson.databind.MapperFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
 
 import com.liferay.commerce.openapi.core.constants.OpenApiPropsKeys;
 
 import java.io.IOException;
-import java.io.OutputStream;
+import java.io.InputStream;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Type;
 
-import javax.ws.rs.Produces;
+import javax.ws.rs.Consumes;
 import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.MultivaluedMap;
-import javax.ws.rs.ext.MessageBodyWriter;
+import javax.ws.rs.ext.MessageBodyReader;
 import javax.ws.rs.ext.Provider;
 
 import org.osgi.service.component.annotations.Component;
@@ -41,49 +39,33 @@ import org.osgi.service.jaxrs.whiteboard.JaxrsWhiteboardConstants;
  */
 @Component(
 	property = {
-		JaxrsWhiteboardConstants.JAX_RS_APPLICATION_SELECT + "=(" + OpenApiPropsKeys.MESSAGE_BODY_WRITERS_ENABLED + ")",
+		JaxrsWhiteboardConstants.JAX_RS_APPLICATION_SELECT + "=(" + OpenApiPropsKeys.MESSAGE_BODY_READERS_ENABLED + ")",
 		JaxrsWhiteboardConstants.JAX_RS_EXTENSION + "=true"
 	},
-	service = MessageBodyWriter.class
+	service = MessageBodyReader.class
 )
-@Produces(MediaType.APPLICATION_JSON)
+@Consumes(MediaType.APPLICATION_JSON)
 @Provider
-public class JsonMessageBodyWriter implements MessageBodyWriter<Object> {
+public class JsonMessageBodyReader implements MessageBodyReader<Object> {
 
-	public long getSize(
-		Object data, Class<?> type, Type genericType, Annotation[] annotations,
-		MediaType mediaType) {
-
-		return -1;
-	}
-
-	public boolean isWriteable(
+	@Override
+	public boolean isReadable(
 		Class<?> type, Type genericType, Annotation[] annotations,
 		MediaType mediaType) {
-
-		if (type == String.class) {
-			return false;
-		}
 
 		return _OBJECT_MAPPER.canSerialize(type);
 	}
 
-	public void writeTo(
-			Object data, Class<?> type, Type genericType,
-			Annotation[] annotations, MediaType mediaType,
-			MultivaluedMap<String, Object> headers, OutputStream out)
+	@Override
+	public Object readFrom(
+			Class<Object> type, Type genericType, Annotation[] annotations,
+			MediaType mediaType, MultivaluedMap<String, String> httpHeaders,
+			InputStream inputStream)
 		throws IOException, WebApplicationException {
 
-		_OBJECT_MAPPER.writeValue(out, data);
-
-		out.flush();
+		return _OBJECT_MAPPER.readValue(inputStream, type);
 	}
 
-	private static final ObjectMapper _OBJECT_MAPPER = new ObjectMapper() {
-		{
-			configure(MapperFeature.SORT_PROPERTIES_ALPHABETICALLY, true);
-			enable(SerializationFeature.INDENT_OUTPUT);
-		}
-	};
+	private static final ObjectMapper _OBJECT_MAPPER = new ObjectMapper();
 
 }
